requirements.txt
django==3.2.*
django-dotenv==1.4.*
djangorestframework==3.12.*
djangorestframework-simplejwt==4.6.*
django-cors-headers==3.7.*
django-csp==3.7.*
pillow==8.2.*
whitenoise==5.2.*
cloudinary==1.25.*
dj-database-url==0.5.*
wheel==0.36.*
psycopg2-binary==2.8.*
gunicorn==20.1.*
flake8==3.9.*
